move types to type
